+++ admin/tabs/preferences/PreferenceModel.js
@@ -4,11 +4,17 @@
  *
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
+import {truncate} from 'lodash';
 import {boolCheckCol, dateTimeCol} from '@xh/hoist/cmp/grid';
 import {HoistModel, LoadSupport, managed} from '@xh/hoist/core';
 import {textArea} from '@xh/hoist/desktop/cmp/input';
-import {RestGridModel} from '@xh/hoist/desktop/cmp/rest';
-import {truncate} from 'lodash';
+import {
+    addAction,
+    bulkDeleteAction,
+    editAction,
+    regroupAction,
+    RestGridModel
+} from '@xh/hoist/desktop/cmp/rest';
 import {DifferModel} from '../../differ/DifferModel';
 
 @HoistModel
@@ -22,6 +28,7 @@ export class PreferenceModel {
         persistWith: this.persistWith,
         enableColChooser: true,
         enableExport: true,
+        selModel: 'multiple',
         store: {
             url: 'rest/preferenceAdmin',
             reloadLookupsOnLoad: true,
@@ -76,6 +83,18 @@ export class PreferenceModel {
         actionWarning: {
             del: 'Are you sure you want to delete? Deleting preferences can break running apps.'
         },
+        toolbarActions: [
+            addAction,
+            editAction,
+            bulkDeleteAction,
+            regroupAction
+        ],
+        menuActions: [
+            addAction,
+            editAction,
+            bulkDeleteAction,
+            regroupAction
+        ],
         columns: [
             {field: 'local', ...boolCheckCol, width: 70},
             {field: 'name', width: 200},