+++ desktop/cmp/rest/data/RestStore.js
@@ -53,6 +53,20 @@ export class RestStore extends UrlStore {
         );
     }
 
+    async bulkDeleteRecordsAsync(records) {
+        const {url} = this,
+            ids = records.map(it => it.id),
+            resp = await XH.fetchJson({
+                url: `${url}/bulkDelete`,
+                params: {ids}
+            }).linkTo(
+                this.loadModel
+            );
+
+        this.loadAsync();
+        return resp;
+    }
+
     async addRecordAsync(rec) {
         return this.saveRecordInternalAsync(rec, true)
             .linkTo(this.loadModel);
@@ -61,6 +75,22 @@ export class RestStore extends UrlStore {
     async saveRecordAsync(rec) {
         return this.saveRecordInternalAsync(rec, false)
             .linkTo(this.loadModel);
+    }
+
+    async regroupRecordsAsync(ids, groupName) {
+        const {url} = this,
+            resp = await XH.fetchService.putJson({
+                url: `${url}/bulkUpdate`,
+                params: {ids},
+                body: {
+                    newParams: {groupName}
+                }
+            }).linkTo(
+                this.loadModel
+            );
+
+        this.loadAsync();
+        return resp;
     }
 
     //--------------------------------